requirements.txt
async-timeout==4.0.2
click==8.1.1
Deprecated==1.2.13
Flask==2.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mimesis==5.3.0
mysql-connector-python==8.0.28
packaging==21.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pyparsing==3.0.7
redis==4.2.1
Werkzeug==2.1.0
wrapt==1.14.0
